+++ pkg/apis/ipam/v1/e2e_shared_test.go
@@ -1,0 +1,3 @@
+package v1_test
+
+const locationIdentifier = "6f43b9222c06479093bb48755afd6dfc" // == ANX82