+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Some examples, more below in the actual changelog (newer entries are more likely
 ### Added
 * api: Add RateLimitError (#438, @nachtjasmin)
 * Replace all occurrences of `vmxnet3` with `virtio` (#444, @nachtjasmin)
+* api: Add search parameter to the prefix api (#456, @kkostial)
 
 ## [0.7.7] -- 2025-01-09